chore: refactor requestHandlers
