+++ packages/extension/src/inpage/inpage.ts
@@ -11,51 +11,18 @@ import {
   typedData,
 } from "starknet"
 
+import { WindowMessageType } from "../shared/MessageType"
+import { getProvider } from "../shared/networks"
+import { EventHandler, StarknetWindowObject } from "./inpage.model"
+import { sendMessage, waitForMsgOfType } from "./messageActions"
 import {
-  AddStarknetChainParameters,
-  EventHandler,
-  StarknetWindowObject,
-} from "./inpage.model"
-import { MessageType, WindowMessageType } from "./shared/MessageType"
-import { getProvider } from "./shared/networks"
+  handleAddNetworkRequest,
+  handleAddTokenRequest,
+} from "./requestMessageHandlers"
 
 const VERSION = `${process.env.VERSION}`
 
-const extId = document
-  .getElementById("argent-x-extension")
-  ?.getAttribute("data-extension-id")
-
 const userEventHandlers: EventHandler[] = []
-
-function sendMessage(msg: MessageType): void {
-  return window.postMessage(
-    { ...msg, extensionId: extId },
-    window.location.origin,
-  )
-}
-
-function waitForMsgOfType<
-  K extends MessageType["type"],
-  T extends { type: K } & MessageType,
->(
-  type: K,
-  timeout: number,
-  predicate: (x: T) => boolean = () => true,
-): Promise<T extends { data: any } ? T["data"] : undefined> {
-  return new Promise((resolve, reject) => {
-    const pid = setTimeout(() => reject(new Error("Timeout")), timeout)
-    const handler = (event: MessageEvent<WindowMessageType>) => {
-      if (event.data.type === type && predicate(event.data as any)) {
-        clearTimeout(pid)
-        window.removeEventListener("message", handler)
-        return resolve(
-          ("data" in event.data ? event.data.data : undefined) as any,
-        )
-      }
-    }
-    window.addEventListener("message", handler)
-  })
-}
 
 // window.ethereum like
 const starknetWindowObject: StarknetWindowObject = {
@@ -66,106 +33,9 @@ const starknetWindowObject: StarknetWindowObject = {
   version: VERSION,
   request: async (call) => {
     if (call.type === "wallet_watchAsset" && call.params.type === "ERC20") {
-      sendMessage({
-        type: "REQUEST_TOKEN",
-        data: {
-          address: call.params.options.address,
-          symbol: call.params.options.symbol,
-          decimals: call.params.options.decimals?.toString(),
-          name: call.params.options.name,
-        },
-      })
-      const { actionHash } = await waitForMsgOfType("REQUEST_TOKEN_RES", 1000)
-
-      if (!actionHash) {
-        // token already exists
-        return false
-      }
-
-      sendMessage({ type: "OPEN_UI" })
-
-      const result = await Promise.race([
-        waitForMsgOfType(
-          "APPROVE_REQUEST_TOKEN",
-          11 * 60 * 1000,
-          (x) => x.data.actionHash === actionHash,
-        ),
-        waitForMsgOfType(
-          "REJECT_REQUEST_TOKEN",
-          10 * 60 * 1000,
-          (x) => x.data.actionHash === actionHash,
-        )
-          .then(() => "error" as const)
-          .catch(() => {
-            sendMessage({ type: "REJECT_REQUEST_TOKEN", data: { actionHash } })
-            return "timeout" as const
-          }),
-      ])
-
-      if (result === "error") {
-        throw Error("User abort")
-      }
-      if (result === "timeout") {
-        throw Error("User action timed out")
-      }
-
-      return true
+      await handleAddTokenRequest(call.params)
     } else if (call.type === "wallet_addStarknetChain") {
-      const callParams = call.params as AddStarknetChainParameters
-
-      sendMessage({
-        type: "REQUEST_CUSTOM_NETWORK",
-        data: {
-          id: callParams.id,
-          name: callParams.chainName,
-          chainId: callParams.chainId,
-          baseUrl: callParams.baseUrl,
-          rpcUrl: callParams.rpcUrl,
-          explorerUrl: callParams.blockExplorerUrl,
-          accountImplementation: callParams.accountImplementation,
-        },
-      })
-
-      const { actionHash } = await waitForMsgOfType(
-        "REQUEST_CUSTOM_NETWORK_RES",
-        1000,
-      )
-
-      if (!actionHash) {
-        return false
-      }
-
-      sendMessage({ type: "OPEN_UI" })
-
-      const result = await Promise.race([
-        waitForMsgOfType(
-          "APPROVE_REQUEST_CUSTOM_NETWORK",
-          11 * 60 * 1000,
-          (x) => x.data.actionHash === actionHash,
-        ),
-        waitForMsgOfType(
-          "REJECT_REQUEST_CUSTOM_NETWORK",
-          10 * 60 * 1000,
-          (x) => x.data.actionHash === actionHash,
-        )
-          .then(() => "error" as const)
-          .catch(() => {
-            sendMessage({
-              type: "REJECT_REQUEST_CUSTOM_NETWORK",
-              data: { actionHash },
-            })
-            return "timeout" as const
-          }),
-      ])
-
-      if (result === "error") {
-        throw Error("User abort")
-      }
-      if (result === "timeout") {
-        throw Error("User action timed out")
-      }
-
-      return true
+      await handleAddNetworkRequest(call.params)
     }
     throw Error("Not implemented")
   },
